Create `Story` struct representing Story API model
First commit. Created a struct `Story` which serializes/deserializes
to the Story API model described here:

https://hackernews.api-docs.io/v0/items/base

I'm not sure what makes the most sense in terms of the API, so I'm
experimenting. These are some questions I've been thinking about:

1. Should I be using `serde_with::double_option` on all my fields? On
the one hand, this ensures that the data structure will parse even with
missing fields, and even if certain fields are keyed to `null`. On the
other hand, maybe it's actually preferable to fail in parsing if certain
fields are missing or none. I think it basically comes down to whether
you want to control the data completeness using the serialization
definition, or using relevant methods on the struct, or its fields.

2. It looks like all the Response models (i.e. Story, Comment, Job, etc.)
share a common base set of fields. Likewise, and aggregate object could
be used to represent all the datasets, and populate which fields are
actually contained in any given response. Not sure if I should simply
define one struct for each, or one general struct, or multiple structs
connected via an enum.