Create CommentNode struct to retrieve a thread
Created a struct called `CommentNode` to create an entire thread. The
problem I'm running into is how to include the hnews::HNClient in to the
mix. One way would be to simply make a CommentNode be generated from a
method on a HNClient instance, and keep a reference back to the client
to be able to hit the API endpoint and get more data.

The main CommentNode's use/benefit I see out of the CommentNode, is the
ability to automatically retrieve a thread of data under a given node.
This is much easier if each Node has a reference to a client. Otherwise,
you have an awkward API and leaky abstraction where you need to pass in
references to an API client all over the place to allow a CommentNode to
retrieve more data.