Prepare to refactor `replies_iter` with references
Attempting one more time to refactor `replies_iter` to use a referance
to the HNClient in the Iterator instance. This will mean that an
HNClient doesn't have to clone itself when creating an Iterator
instance. I think using a queue of simply Ids will make this a little
easier and more successful.