Disable `cargo test` in github workflow
Disable `cargo test` in the github workflow. There are some unit tests
that depend on test data files, or environment variables which are not
present when running the github CI. I have not yet figured out how I'll
handle this, as it would be nice to have all valid tests to run in the
CI.

For now I'll just run `cargo test` locally.