Begin model unification branch
This commit is essentially just providing a difference to git, so that I
can open a PR branch in GitHub. I'm getting in to the habit of opening a
draft PR when I open a branch, and documenting what the goals/purpose of
the branch is.

The actual commit is just deleting a comment, which was a TODO I hope to
resolve over this PR.