Fixed dead comments issue (which was actually different)
So the problem was not dead comments, which is what I originally
assumed. The comments parsing process begins with querying the HTML
document for the root of the comment table. However, if there are no
comments, then there is no comment table, and so the query yields no
results.

I have resolved the issue by adding a check as to whether the query for
the comment root retrieved any results.