Added new `Thread` model type, refactor comment logic
Added a new model type called `Thread`. A thread is essentially just the
combination of a Listing and a Comment tree, where the Comment tree is a
Vec<Comment> that has already been nested appropriately.

One thing I realized is that currently I don't have a way to get the
text associated with an original post. For example with a HackerNews
post, I can't get the original question text. I think this should be a
field on the Listing type, probably as an Option<String> to also cover
situations where this text is not present.

However, the Listing HTML is kind of not-so-convenient to work with.
Right now, I'm using a `next_sibling()` method to derive where a certain
node is, as opposed to querying based on CSS/HTML descriptors. However,
I think I can refactor to make this a little better. Instead of querying
for the title node via 'tr.athing:not(.comtr)', I can query for the
parent node via `table.fatitem`. I'll still probably need to use some
`next_sibling()` method calls, but hopefully by starting at the parent I
can assess the overall number of rows present to assess whether the
listing text is present or not.