Progress on listings parser addition for comment
Made progress on additions to listings parser. Now have logic to assess
whether a listing contains a comment node or not. Need to refactor a
little bit, but should be able to use most of the logic as is. Instead
of doing a ListingType and match on the variants, I'm just going to have
a single method `parse_post_comment` and return an Option. For listings
without a comment, I'll just return None.