Add comment text as optional, add dead flag
Added in a dead flag to the comment data structure definition, and added
in logic to parse out whether a comment is dead. Also since dead
comments by default show up as "[flagged]" with no user text, when a
comment is recognized as dead, I am not parsing for the user comment
text.

Potentially this should be updated in the future if I add in user
authentication, which adds a setting to show dead comments.