Progress on improving error handling
This commit changes the trait interface for `HnCommand`. It required
that when the `cmd` function returns, the Error type should be a
concrete `HnError` instance. In other words, this means that when we
make it up to the `cmd` function in the call stack, all underlying
errors should be mapped to the concrete error type of the application.

My idea is that this error type will have augmented formatted error
printing capabilities akin to what cargo does.

Possible followup of actually moving the CLI into it's own repository.
Still have mixed feelings on this, so not sure if I will do it.