More work on error formatting
Adding better log statements, specifically around cli commands. I want
to easily tell which subcommand's code path we went down, and if there
is a failure, where about in the code it occurred.

This commit currently sets an invalid url in the `fn thread()` method of
the HtmlClient. I am making this commit because I am about to apply
`cargo fmt`.